Feature/test harness (#4779)
* refactoring browser tests

* refactoring browser tests

* refactoring browser tests

* fix configure file loading

* fixing configure loading

+++ not_for_release/testFramework/Browser/Traits/ConfigureFileConcerns.php
@@ -9,17 +9,20 @@ trait ConfigureFileConcerns
 
     protected function createInitialConfigures()
     {
+        $configFile = 'NOT FOUND';
         $this->saveConfigures(DIR_FS_ROOT);
-        if (file_exists($configFile = DIR_FS_ROOT . 'not_for_release/testFramework/Browser/zencartConfigures/admin.' . $this->user . '.configure.php')) {
-            copy($configFile, DIR_FS_ROOT . 'admin/includes/configure.php');
-        } elseif (file_exists($configFile = DIR_FS_ROOT . 'not_for_release/testFramework/Browser/zencartConfigures/admin.default.configure.php')) {
+        if (file_exists($configFile = DIR_FS_ROOT . '/not_for_release/testFramework/Browser/zencartConfigures/local/admin.' . $this->user . '.configure.php')) {
+            copy($configFile, DIR_FS_ROOT . '/admin/includes/configure.php');
+        } elseif (file_exists($configFile = DIR_FS_ROOT . '/not_for_release/testFramework/Browser/zencartConfigures/admin.default.configure.php')) {
             copy($configFile, DIR_FS_ROOT . 'admin/includes/configure.php');
         }
-        if (file_exists($configFile = DIR_FS_ROOT . 'not_for_release/testFramework/Browser/zencartConfigures/catalog.' . $this->user . '.configure.php')) {
-            copy($configFile, DIR_FS_ROOT . 'includes/configure.php');
-        } elseif (file_exists($configFile = DIR_FS_ROOT . 'not_for_release/testFramework/Browser/zencartConfigures/catalog.default.configure.php')) {
-            copy($configFile, DIR_FS_ROOT . 'includes/configure.php');
-        }
+        echo "admin config fle = " . $configFile . "\n";
+//        if (file_exists($configFile = DIR_FS_ROOT . '/not_for_release/testFramework/Browser/zencartConfigures/local/catalog.' . $this->user . '.configure.php')) {
+//            copy($configFile, DIR_FS_ROOT . '/includes/configure.php');
+//        } elseif (file_exists($configFile = DIR_FS_ROOT . '/not_for_release/testFramework/Browser/zencartConfigures/catalog.default.configure.php')) {
+//            copy($configFile, DIR_FS_ROOT . 'includes/configure.php');
+//        }
+//        echo "store config fle = " . $configFile . "\n";
     }
 
     public function copyFile($fileToRemove)
@@ -37,14 +40,18 @@ trait ConfigureFileConcerns
     {
         $this->saveConfigures($rootPath);
         $file = $rootPath . '/includes/configure.php';
-        unlink($file);
-        touch($file);
-        chmod($file, 0777);
+        if (file_exists($file)) {
+            unlink($file);
+            touch($file);
+            chmod($file, 0777);
+        }
 
         $file = $rootPath . '/admin/includes/configure.php';
-        unlink($file);
-        touch($file);
-        chmod($file, 0777);
+        if (file_exists($file)) {
+            unlink($file);
+            touch($file);
+            chmod($file, 0777);
+        }
     }
 
     protected function saveConfigures($rootPath)
@@ -57,7 +64,9 @@ trait ConfigureFileConcerns
 
     protected function saveConfigureFile($dest)
     {
-        copy($dest, $dest.'.config.save');
+        if (file_exists($dest)) {
+            copy($dest, $dest . '.config.save');
+        }
     }
 
 }
